requirements.txt
msal>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.7.0
typer>=0.9.0
pandas>=2.1.0
openpyxl>=3.1.0
customtkinter>=5.2.0
CTkMessagebox>=2.5
CTkListbox>=1.3
